Set-Up Google oAuth

+++ controllers/api/v1/post_api.js
@@ -22,13 +22,20 @@ module.exports.destroy = async function (req, res) {
     try{
         let post=await Post.findById(req.params.id);
           
-      
+            if(post.user==req.user.id)
+            {
                 post.remove();
            
                 await Comment.deleteMany({post:req.params.id});
                 return res.json(200,{
                     message:"Deleted the post"
                 })
+            }
+            else{
+                return res.josn(401,{
+                    message:'You cannot delete this post'
+                })
+            }
         }
     catch(err){
         // req.flash('error',err);
